chore: refactor agent observer
